Example projects fixed (#22)
* Fix skipToBytes self-overlap misses with KMP search

* Document that deleteMetadata requires a readable file

* Fix isMetadataEmbeddable to include GIF

* Fix Int overflow in TIFF image size directory selection

* Fix minor project-rule and dead-code violations

Correct the RAF header skip message to name the constant value in all
three RAF files, collapse dead conditionals in PngImageParser, and
convert the JxlReader line comment to a block comment.

* Fix silent MakerNote parse-failure skip in ExifToolDumpComparisonTest

A dump MakerNote section with comparable values but a NULL Kim
directory is now a recorded failure instead of being skipped.
Empty sections (ExifTool warning only) remain skipped.

* Publish kotlinx-io, ktor-io and kotlinx-datetime as api dependencies

Their types (TimeZone, Path, Source, ByteReadChannel) appear in
public Kim signatures, so consumers could not compile against the
documented API without declaring them themselves.

* Document the runBlocking deadlock constraint of KtorByteReadChannelByteReader

The reader must not be used from the dispatcher that serves the
channel, and its unsynchronized buffer state confines it to one
thread. Both constraints are now stated in the class KDoc and on
both KimKtor.readMetadata(ByteReadChannel, ...) overloads.

* Fix example projects and cover them with CI

Bump both example projects from kim 0.32.0 to 0.38.1 so TiffWriter and
kotlinx-io APIs resolve, add explicit kotlinx-io-core dependency to the
Kotlin sample, fix unclosed JvmInputStreamByteReader leaks in
updateTakenDateLowLevelApi and setGeoTiffToJpeg by wrapping the reader
in use, fix Java sample resource handling and Gradle 9 compatibility
by guarding kotlin platform attribute to resolvable configurations,
upgrade both Gradle wrappers from 8.7 to 9.7.1 (jar, properties,
gradlew scripts), and add CI steps that build each example so they
cannot rot.

+++ src/commonMain/kotlin/de/stefan_oltmann/kim/format/tiff/TiffImageParser.kt
@@ -214,7 +214,10 @@ public object TiffImageParser : ImageParser {
         for (directory in tiffContents.directories)
             getImageSize(directory)?.let { imageSizes.add(it) }
 
-        return imageSizes.maxByOrNull { it.width * it.height }
+        /*
+         * Long arithmetic, so areas of huge images cannot overflow Int.
+         */
+        return imageSizes.maxByOrNull { it.width.toLong() * it.height }
     }
 
     private fun getImageSize(directory: TiffDirectory): ImageSize? {
